(global) move harnesses into shared testing packages #6
there are no protractor dependencies, so they can be used in protractor as well as in testbed
