+++ libs/core/testing/core.component-harness.ts
@@ -1,4 +1,4 @@
-import { BaseComponentHarness } from './base.co';
+import { BaseComponentHarness } from './base-component-harness';
 
 export class CoreComponentHarness extends BaseComponentHarness {
   static hostSelector = 'company-core';
