+++ libs/core/testing/index.ts
@@ -1,0 +1,2 @@
+export * from './base-component-harness';
+export * from './core.component-harness';
