+++ libs/core/testing/ng-package.json
@@ -1,0 +1,7 @@
+{
+  "$schema": "../../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../../dist/company/core/testing",
+  "lib": {
+    "entryFile": "index.ts"
+  }
+}
