+++ libs/core/testing/protractor/index.ts
@@ -1,5 +1,3 @@
-export * from './base.co';
-export * from './core.co';
 export * from './expect.functions';
 export * from './init.functions';
 export * from './spec-shot.functions';
