+++ libs/hero/testing/index.ts
@@ -1,0 +1,1 @@
+export * from './hero.component-harness';
