+++ libs/hero/testing/ng-package.json
@@ -1,0 +1,10 @@
+{
+  "$schema": "../../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../../dist/company/hero/testing",
+  "lib": {
+    "entryFile": "index.ts",
+    "umdModuleIds": {
+      "@company/core/testing": "@company/core/testing"
+    }
+  }
+}
