--- libs/hero/testing/protractor/ng-package.json
@@ -1,11 +1,0 @@
-{
-  "$schema": "../../../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../../../dist/company/core/testing/protractor",
-  "lib": {
-    "entryFile": "index.ts",
-    "umdModuleIds": {
-      "protractor": "protractor",
-      "@company/core/testing/protractor": "@company/core/testing/protractor"
-    }
-  }
-}
